move more common stuff out of RunningStitch
