+++ lib/sew_stack/stitch_layers/running_stitch/running_stitch_layer.py
@@ -1,7 +1,4 @@
-from copy import copy
-
 from ....i18n import _
-from ....stitch_plan import StitchGroup
 from ....stitches.running_stitch import running_stitch
 from ....svg import PIXELS_PER_MM
 from ....utils.classproperty import classproperty
@@ -31,20 +28,6 @@ class RunningStitchLayerEditor(StitchLayerEditor, RandomizationPropertiesMixin, 
                          min=0.01,
                          unit="mm",
                          ),
-                Category(_("Repeats")).children(
-                    Property(
-                        "repeats", _("Repeats"),
-                        help=_('Defines how many times to run down and back along the path.'),
-                        type=int,
-                        min=1,
-                    ),
-                    Property(
-                        "repeat_stitches", _("Repeat exact stitches"),
-                        type=bool,
-                        help=_('Should the exact same stitches be repeated in each pass?  ' +
-                               'If unchecked, different randomization settings are applied on each pass.'),
-                    ),
-                ),
                 cls.path_properties(),
             ),
             cls.randomization_properties().children(
@@ -69,8 +52,6 @@ class RunningStitchLayer(StitchLayer, RandomizationMixin, PathMixin):
             stitch_length=2,
             tolerance=0.2,
             stitch_length_jitter_percent=0,
-            repeats=1,
-            repeat_stitches=True,
         )
         defaults.update(cls.randomization_defaults())
         defaults.update(cls.path_defaults())
@@ -81,46 +62,15 @@ class RunningStitchLayer(StitchLayer, RandomizationMixin, PathMixin):
     def layer_type_name(self):
         return _("Running Stitch")
 
-    def get_num_copies(self):
-        if self.config.repeat_stitches:
-            # When repeating stitches, we use multiple copies of one pass of running stitch
-            return 1
-        else:
-            # Otherwise, we re-run running stitch every time, so that
-            # randomization is different every pass
-            return self.config.repeats
-
     def running_stitch(self, path):
-        stitches = []
-
-        for i in range(self.get_num_copies()):
-            if i % 2 == 0:
-                this_path = path
-            else:
-                this_path = list(reversed(path))
-
-            stitches.extend(running_stitch(
-                this_path,
+        return running_stitch(
+                path,
                 self.config.stitch_length * PIXELS_PER_MM,
                 self.config.tolerance * PIXELS_PER_MM,
                 (self.config.stitch_length_jitter_percent > 0),
                 self.config.stitch_length_jitter_percent,
                 self.get_random_seed()
-            ))
-
-        self.offset_stitches(stitches)
-
-        if self.config.repeats > 0 and self.config.repeat_stitches:
-            repeated_stitches = []
-            for i in range(self.config.repeats):
-                if i % 2 == 0:
-                    repeated_stitches.extend(copy(stitches))
-                else:
-                    # reverse every other pass
-                    repeated_stitches.extend(reversed(copy(stitches)))
-            stitches = repeated_stitches
-
-        return StitchGroup(stitches=stitches, color=self.stroke_color)
+            )
 
     def to_stitch_groups(self, previous_stitch_group, next_element):
-        return [self.running_stitch(path) for path in self.get_paths()]
+        return self.stitch_paths(self.running_stitch, self.apply_random_stitch_offset)
